Fixed some warnings

+++ src/server/game/Spells/SpellEffects.cpp
@@ -7679,7 +7679,7 @@ void Spell::EffectCastButtons(SpellEffIndex effIndex)
             continue;
 
         uint32 cost = spellInfo->CalcPowerCost(m_caster, spellInfo->GetSchoolMask());
-        if (m_caster->GetPower(POWER_MANA) < cost)
+        if (uint32(m_caster->GetPower(POWER_MANA)) < cost)
             continue;
 
         TriggerCastFlags triggerFlags = TriggerCastFlags(TRIGGERED_IGNORE_GCD | TRIGGERED_IGNORE_CAST_IN_PROGRESS | TRIGGERED_CAST_DIRECTLY);
